+++ src/fr/maxlego08/menu/loader/actions/MessageLoader.java
@@ -3,13 +3,11 @@ package fr.maxlego08.menu.loader.actions;
 import fr.maxlego08.menu.api.loader.ActionLoader;
 import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.utils.TypedMapAccessor;
-import fr.maxlego08.menu.requirement.actions.BroadcastAction;
 import fr.maxlego08.menu.requirement.actions.MessageAction;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 public class MessageLoader implements ActionLoader {
 
@@ -21,7 +19,17 @@ public class MessageLoader implements ActionLoader {
     @Override
     public Action load(String path, TypedMapAccessor accessor, File file) {
         boolean miniMessage = accessor.getBoolean("minimessage", accessor.getBoolean("mini-message", true));
-        List<String> messages = accessor.getStringList("messages");
+        List<String> messages = new ArrayList<>();
+        if (accessor.contains("message")) {
+            messages.add(accessor.getString("message"));
+        } else if (accessor.contains("messages")) {
+            Object element = accessor.getObject("messages", new ArrayList<>());
+            if (element instanceof String) {
+                messages.add((String) element);
+            } else {
+                messages = accessor.getStringList("messages");
+            }
+        }
         return new MessageAction(messages, miniMessage);
     }
 }