+++ src/fr/maxlego08/menu/loader/materials/MagicCosmeticsLoader.java
@@ -1,6 +1,7 @@
 package fr.maxlego08.menu.loader.materials;
 
 import fr.maxlego08.menu.api.loader.MaterialLoader;
+import org.bukkit.Material;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -15,7 +16,10 @@ public class MagicCosmeticsLoader implements MaterialLoader {
 
     @Override
     public ItemStack load(Player player, YamlConfiguration configuration, String path, String materialString) {
-        return MagicAPI.getEquipped(player.getName(), materialString);
-
+        ItemStack itemStack = MagicAPI.getEquipped(player.getName(), materialString);
+        if (itemStack == null){
+            return new ItemStack(Material.AIR);
+        }
+        return itemStack;
     }
 }