+++ src/fr/maxlego08/menu/zcore/utils/discord/ReturnConsumer.java
@@ -1,0 +1,8 @@
+package fr.maxlego08.menu.zcore.utils.discord;
+
+@FunctionalInterface
+public interface ReturnConsumer<T, G> {
+
+	G accept(T t);
+	
+}